requirements.txt
matplotlib==3.1.0
numba==0.44.1
pytorch_toolbelt==0.4.1
torch==1.7.1
tqdm==4.32.1
albumentations==0.4.2
pandas==0.24.2
opencv_python==4.5.1.48
torchvision==0.8.2
simplecv==0.3.4
numpy==1.20.1
scikit_image==0.18.1
segmentation_models_pytorch==0.1.3
osgeo==0.0.0
Pillow==8.1.2
skimage==0.0
glob2==0.7
pathlib2==2.3.4